+++ src/allotropy/parsers/agilent_gen5/constants.py
@@ -1,8 +1,11 @@
 from enum import Enum
 
 MULTIPLATE_FILE_ERROR = "Only a single plate per file can be processed at this time. Please refer to Gen5 documentation for how to generate single plate exports from multi-plate experiments"
+MULTIPLE_READ_MODE_ERROR = "Only a single endpoint read per file can be processed at this time. Please refer to the Gen5 documentation for supported exports"
 NO_PLATE_DATA_ERROR = "No plate data found in file."
-UNSUPORTED_READ_TYPE_ERROR = "Only Endpoint measurements can be processed at this time."
+UNSUPPORTED_READ_TYPE_ERROR = (
+    "Only Endpoint measurements can be processed at this time."
+)
 
 HEADER_PREFIXES = frozenset(
     {
@@ -43,6 +46,11 @@ class ReadMode(str, Enum):
     FLUORESCENCE = "Fluorescence"
     LUMINESCENCE = "Luminescence"
     ALPHALISA = "Alpha"
+
+
+UNSUPPORTED_READ_MODE_ERROR = (
+    f"Read mode not found; expected to find one of {sorted(ReadMode._member_names_)}."
+)
 
 
 class ReadType(str, Enum):